requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
python-dotenv==0.12.0
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
